fix some imports from refactor dirs
